requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
openai==1.3.0
PyPDF2==3.0.1
python-docx==0.8.11
psycopg2-binary==2.9.9
pytest==7.4.3
httpx==0.25.1
pdfminer.six==20221105
pandas==2.1.1
numpy==1.26.1
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
python-magic==0.4.27
aiohttp==3.8.6
email-validator==2.0.0